+++ CHANGELOG.md
@@ -19,6 +19,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `KipperLogger.levels` as a static variable to access the enum `LogLevel`.
 - New abstract base class `KipperParseToken`, which will represent the major parse tokens inside a kipper program. 
   This class will be used to represent the items in a kipper program, and generate the typescript counterpart.
+  Current child classes:
+  - `Expression`
+  - `FunctionDefinition`
+  - `Declaration`
+  - `CompoundStatement`
+  - `SelectionStatement`
+  - `ExpressionStatement`
+  - `IterationStatement`
+  - `JumpStatement`
 - Properties `parser`, `lexer`, `errorHandler` and `tokenStream` inside the class `KipperFileContext`.
 - New dictionary `builtInFunctions`, which will map to the built-in function names, like `print()`, a 
   handler that will generate the typescript code to replace the function call with. This is to avoid having to
@@ -31,7 +40,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed bug in `KipperErrorListener`, which resulted in errors being not properly raised. The function is
   now a template, where `<T>` represents the offending token. `<T>` will also be passed onto `KipperSyntaxError<T>`.
 - Changed type of `LogLevel`, which now returns string-representations of the log level.
-
+- Fixed `initializer` rule in Kipper.g4 and removed invalid designator rules.
+- 
 ### Removed
 - Unneeded namespace `KipperStreams` and its functions.
 - Unneeded variable `LogLevelNames`, as now the enum `LogLevel` per default returns the names of the level.